chore: consolidate typings
