+++ src/frontend/hooks/useTabNavigationStore.ts
@@ -1,5 +1,5 @@
 import {create} from 'zustand';
-import {TabName} from '../Navigation/types';
+import {TabName} from '../sharedTypes';
 
 type NavigationStoreState = {
   currentTab: TabName;
